Finish concurrent ticket buying
when a seat selected ajax request sent to server and temporary store the
seat in block list. When another request came to the same seat since it
is already there in seat block list user will get a message to try
another seat. When user exit from the page all seats in block list for
the above user will be cleared. So any new user can not see the seats
selected but not paid by first user after the first user exit from the
page.

When click buy or print the same seat cleaning from block list will
happen and redirect to the payment method with the selected seats. After
the seat got paid it will visible to other users as reserved seats.